chore: refactor, move documents to own folder
