+++ tenant/src/components/documents/Identification.vue
@@ -67,7 +67,7 @@
 <script lang="ts">
 import { Component, Vue } from "vue-property-decorator";
 import { mapState } from "vuex";
-import DocumentInsert from "@/components/DocumentInsert.vue";
+import DocumentInsert from "@/components/documents/DocumentInsert.vue";
 import FileUpload from "@/components/uploads/FileUpload.vue";
 import { DocumentType } from "df-shared/src/models/Document";
 import { UploadStatus } from "@/components/uploads/UploadStatus";
